- Improve aclocal check in autogen.sh - Fix flex binary detection in configure.ac

+++ autogen.sh
@@ -6,10 +6,11 @@
 ## autoconf 2.x
 TESTAUTOHEADER="autoheader autoheader-2.53 autoheader2.50 autoheader259 autoheader253"
 TESTAUTOCONF="autoconf autoconf-2.53 autoconf2.50 autoconf259 autoconf253"
+TESTACLOCAL="aclocal aclocal19"
 
 AUTOHEADERFOUND="0"
 AUTOCONFFOUND="0"
-
+ACLOCALFOUND="0"
 
 ##
 ## Look for autoheader
@@ -38,6 +39,17 @@ for i in $TESTAUTOCONF; do
         fi
 done
 
+##
+## Look for aclocal
+##
+for i in $TESTACLOCAL; do
+        if which $i > /dev/null 2>&1; then
+                ACLOCAL=$i              
+                ACLOCALFOUND="1"
+                break
+        fi
+done
+
 
 ##
 ## do we have it?
@@ -47,10 +59,17 @@ if test "$AUTOCONFFOUND" = "0" -o "$AUTOHEADERFOUND" = "0"; then
         exit 1
 fi
 
+if test "$ACLOCALFOUND" = "0"; then
+        echo "$0: aclocal not found" >&2
+        exit 1
+fi
+
+
 rm -rf autom4te*.cache
 rm -f configure include/config.h*
 
-aclocal
+echo "$0: running $ACLOCAL"
+$ACLOCAL || exit 1
 
 echo "$0: running $AUTOHEADER"
 $AUTOHEADER || exit 1
